perbaikan pelanggan dan connect ke server

+++ server/src/main/resources/database/schema.sql
@@ -34,7 +34,7 @@ CREATE TABLE menu_items (
     menu_id SERIAL PRIMARY KEY,
     name VARCHAR(100) NOT NULL,
     description TEXT,
-    price DECIMAL(10, 2) NOT NULL,
+    price INT NOT NULL,
     category menu_category NOT NULL,
     image_path VARCHAR(255),
     is_available BOOLEAN DEFAULT TRUE
